Minor touches to playerspawning
