+++ Assets/Scripts/Scripts/Working Scripts/InfinigunItem.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 74c90ab388ee5c24fbbcf4649e7f407f
+timeCreated: 1507226244
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
